requirements.txt
Django==1.10.3
psycopg2==2.6.2
django-fsm==2.4.0
djangorestframework==3.5.3
factory_boy==2.7.0